refactor: rename category api file name
