+++ petbook_fe/lib/utils/gerHrefWithCategory.ts
@@ -1,4 +1,4 @@
-import { CategoryItem } from "@lib/API/petBookAPI/types/categoryRequestSpr";
+import { CategoryItem } from "@lib/API/petBookAPI/types/categoryRequest";
 
 export default function getHrefWithCategory(category: CategoryItem) {
   return `/community/list?category=${category.name ? category.name : ""}_${
